backend/requirements.txt
fastapi~=0.115.8
uvicorn~=0.34.0
requests~=2.32.3
pymongo~=4.11.1

python-dotenv==1.0.1
schedule~=1.2.2

openai~=1.62.0
azure-ai-vision-imageanalysis~=1.0.0
azure-identity~=1.20.0
ffmpeg-python~=0.2.0
ffmpegcv~=0.3.16

langchain-text-splitters~=0.3.6
langchain_community~=0.3.17
langchain_openai~=0.3.6

datasets~=3.3.1
ragas~=0.2.13
bcrypt~=4.3.0
